requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-mirage-field==1.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
exrex==0.10.5
flake8==3.9.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
